Find the Duplicate Number.cpp
You are given an array arr containing n+1 integers, where each integer is in the range [1, n] inclusive. There is exactly one duplicate number in the array, but it may appear more than once. Your task is to find the duplicate number without modifying the array and using constant extra space.

Input :
An integer array arr of size n+1, where each element is an integer in the range [1, n].
Example : arr = [3, 1, 3, 4, 2]

Output :
Return the duplicate integer present in the array.
Example: Duplicate number: 3

+++ Day3.cpp
@@ -8,7 +8,7 @@ int main() {
     cout<<"Enter a n:";
     cin>>n;
     
-    int arr[n+1];
+    long long arr[n+1];
     cout<<"Enter n+1 element from 1 to n:"<<endl;
     for(int i=0; i<=n; i++) {
         cout<<"Element "<<i<<":";
@@ -17,8 +17,8 @@ int main() {
     
     // *** MAIN CODE ***
     
-    int slow = arr[0];
-    int fast = arr[0];
+    long long slow = arr[0];
+    long long fast = arr[0];
     
     slow = arr[slow];
     fast = arr[arr[fast]];
